// apps/mobile/src/api/generated/kitchenCreate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hyperlocal Kitchen API
 * OpenAPI spec version: 0.1.0
 */

export interface KitchenCreate {
  zone_id?: string | null;
  address_label?: string | null;
  community_id: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  name: string;
  description?: string | null;
  pickup_enabled?: boolean;
  delivery_enabled?: boolean;
  /**
     * @minimum 0
     * @maximum 1000000
     */
  delivery_fee_paise?: number;
  upi_id?: string | null;
  fssai_number?: string | null;
}