// apps/mobile/src/api/generated/listingUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hyperlocal Kitchen API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingUpdateStatus } from './listingUpdateStatus';

export interface ListingUpdate {
  pickup_point_ids?: string[] | null;
  service_date?: string | null;
  available_from?: string | null;
  available_until?: string | null;
  order_cutoff?: string | null;
  price_paise?: number | null;
  quantity_total?: number | null;
  pickup_enabled?: boolean | null;
  delivery_enabled?: boolean | null;
  status?: ListingUpdateStatus;
}